API tweaks
